Rename ppo_policy into ppo_tf_policy.
